Add last current weather

+++ src/main/java/fr/brde/api/repositories/CurrentWeatherRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CurrentWeatherRepository extends JpaRepository<CurrentWeather, Integer>
+public interface CurrentWeatherRepository extends JpaRepository<CurrentWeather, Long>
 {
 
 }
